musl licenses, and signed warnings

+++ source/algos/musl1.c
@@ -2,25 +2,15 @@
  * SMART: string matching algorithms research tool.
  * Copyright (C) 2012  Simone Faro and Thierry Lecroq
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>
- *
- * contact the authors at: faro@dmi.unict.it, thierry.lecroq@univ-rouen.fr
- * download the tool at: http://www.dmi.unict.it/~faro/smart/
- *
+ * The musl memmem implementation is licensed under
+ * the standard MIT license.
+ * Copyright © 2005-2020 Rich Felker, et al.
  */
 
 #include "include/define.h"
 #include "include/main.h"
 
+#include <stddef.h>
 #include <stdint.h>
 
 static char *twobyte_memmem(const unsigned char *h, size_t k,
@@ -130,7 +120,7 @@ static char *twoway_memmem(const unsigned char *h, const unsigned char *z,
   /* Search loop */
   for (;;) {
     /* If remainder of haystack is shorter than needle, done */
-    if (z - h < l)
+    if (z - h < (ptrdiff_t)l)
       return 0;
 
     /* Check last byte first; advance by shift on mismatch */
